Fix failing tests
At the time of this fix, these tests were passing on Travis, however,
failed on my Mac locally running on node v12.18.1.

After a little digging, the underlying error seemed to be related to
schema validation issues, occurring when calling validateSchema
function within app.ready(...).
It seems to be a valid error that happens when attempting to initialize
Graphql with invalid schema, however, I'm unsure why I encountered it
and not other contributors (including CI).