Expose the generated GraphQLSchema object for alternative uses
`fastify-gql` currently keeps the schema it executes queries against private. If you want to execute GraphQL queries against your `fastify-gql` managed schema in another way, like in tests, from background jobs, or using your own protocol bindings of some sort, this is unfortunate! I think it makes sense to expose it so developers can access it and do whatever they need to with it. In the normal course of writing GraphQL apps this shouldn't be necessary, but it doesn't seem like `fastify-gql` needs to reserve the schema as an implementation detail since the type is already in use outside `fastify-gql`'s API boundaries.