Update subscriptions.md
Custom PubSub documentation is diverged from the runtime. The value of close is expected to be an array at runtime.  Another option would be to alter the lib/execution to manage either an array or a function. I can supply a PR if that is the desired change. This proposed change addresses a documentation update.